refactor: move puddle transfers to shared
