+++ Content.Shared/Fluids/SharedPuddleSystem.Spillable.cs
@@ -21,7 +21,6 @@ namespace Content.Shared.Fluids;
 
 public abstract partial class SharedPuddleSystem
 {
-    [Dependency] protected readonly OpenableSystem Openable = default!;
 
     protected virtual void InitializeSpillable()
     {
